Release 0.1.0

+++ pool_boy.sh
@@ -3,8 +3,18 @@
 set -e
 set -x
 
+PRIVATE_KEY=/root/.ssh/id_rsa
+
 echo
 echo "Adding git SSH private key"
+
+if [ -e $PRIVATE_KEY ]
+then
+    echo "ERROR Private key $PRIVATE_KEY already existing."
+    echo "You should run this script only from a run-once Docker container for Concourse"
+    exit 1
+fi
+
 set +x
 echo "${POOL_REPO_SSH_PRIVATE_KEY}" > /root/.ssh/id_rsa
 set -x
